requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.41.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
contourpy==1.1.1
cycler==0.12.1
Eel==0.16.0
fonttools==4.43.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.7.3
mpmath==1.3.0
numpy==1.24.4
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pycparser==2.21
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pywin32-ctypes==0.2.2
six==1.16.0
sympy==1.12
tk==0.1.0
whichcraft==0.6.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
